refactor: Move checkauthentication to api folder
