+++ client/src/Api/checkAuthentication.js
