+++ client/src/Api/index.js
@@ -1,7 +1,9 @@
 import deleteArticleFetch from "./deleteArticleFetch"
 import deleteUserFetch from "./deleteUserFetch"
+import checkAuthentication from "./checkAuthentication"
 
 export {
     deleteArticleFetch,
     deleteUserFetch,
+    checkAuthentication,
 }
